// src/contracts/enums.ts
export type Verbosity = number;

export enum BackgroundColors {
  error = '#dc3545',
  info = '#6c757d',
  notice = '#17a2b8',
  success = '#28a745',
  warn = '#ffc107',
}

export enum ForegroundColors {
  error = '#ffffff',
  info = '#ffffff',
  notice = '#ffffff',
  success = '#ffffff',
  warn = '#ffffff',
}